Setup Flashcard table view cell

+++ Flashcards/Controllers/Decks/DecksListViewController.swift
@@ -61,7 +61,7 @@ class DecksListViewController: UITableViewController {
     }
     
     override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        CGFloat(80)
+        80
     }
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
